+++ packages/emnapi/include/emnapi_common.h
